+++ R/data-worldbank.R
@@ -8,6 +8,7 @@
 #' @format JSON
 #' @examples
 #'
+#' \dontrun{
 #' library(dplyr)
 #'
 #' # worldbank is a 500 length character vector of JSON
@@ -36,4 +37,5 @@
 #'   count(sector) %>%              # Count by sector
 #'   arrange(desc(n)) %>%           # Arrange descending
 #'   top_n(10)                      # Take the top 10
+#' }
 NULL