+++ tests/testthat/test-examples.R
@@ -1,0 +1,3 @@
+context("examples")
+
+test_examples(rprojroot::find_testthat_root_file("../.."))