+++ tests/testthat/test-json_structure.R
@@ -23,7 +23,6 @@ test_that("simple string works", {
 })
 
 test_that("simple object works", {
-  skip("wait: testthat")
   actual <- '{"name": "value"}' %>% json_structure
   expected <- tbl_json(
       dplyr::tibble(
@@ -73,7 +72,6 @@ test_that("simple array works", {
 
 test_that("nested object works", {
 
-  skip("wait: testthat")
   expect_identical(
     '{"k1": {"k2": "value"}}' %>% json_structure,
     tbl_json(